% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.tests.R
\name{edge.tests}
\alias{edge.tests}
\title{This function does the tests by Chen and Friedman}
\usage{
edge.tests(x, y)
}
\arguments{
\item{x}{a matrix of numbers}

\item{y}{a matrix of numbers}
}
\value{
a list with statistics and p values
}
\description{
This function does the tests by Chen and Friedman
}
