% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfun.R
\name{mainfun}
\alias{mainfun}
\title{just a test function}
\usage{
mainfun(dta)
}
\arguments{
\item{dta}{a list}
}
\value{
object of type test
}
\description{
just a test function
}
\examples{
dta=list(
  x=matrix(stats::runif(200), 100, 2),
  y=matrix(stats::runif(300), 150, 2)
)
mainfun(dta)

}
